# swagger.yml
swagger: '2.0'
info:
  description: This is an API for accessing all sorts of college football data.  Please note that API keys should be supplied with "Bearer " prepended (e.g. "Bearer your_key"). API keys can be acquired from the CollegeFootballData.com website.
  version: 4.3.2
  title: College Football Data API
  contact:
    email: admin@collegefootballdata.com
host: api.collegefootballdata.com
basePath: /
...
schemes:
  - https
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
security:
  - ApiKeyAuth: []
paths:
...
  /teams:
    get:
      tags:
        - teams
      summary: Team information
      description: Get team information
      operationId: getTeams
      produces:
        - application/json
      parameters:
        - name: conference
          description: Conference abbreviation filter
          in: query
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Team'
        '400':
          description: error
...
  /player/search:
    get:
      tags:
        - players
      summary: Search for player information
      description: Search for players
      operationId: playerSearch
      produces:
        - application/json
      parameters:
        - name: searchTerm
          description: Term to search on
          in: query
          required: true
          type: string
        - name: position
          description: Position abbreviation filter
          in: query
          required: false
          type: string
        - name: team
          description: Team filter
          in: query
          required: false
          type: string
        - name: year
          description: Year filter
          in: query
          required: false
          type: integer
          minimum: 2001
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/PlayerSearchResult'
        '400':
          description: error
  /player/usage:
    get:
      tags:
        - players
      summary: Player usage metrics broken down by season
      description: Player usage metrics by season
      operationId: getPlayerUsage
      produces:
        - application/json
      parameters:
        - name: year
          description: Year filter
          in: query
          required: true
          default: 2020
          type: integer
          minimum: 2013
        - name: team
          description: Team filter
          in: query
          required: false
          type: string
        - name: conference
          description: Conference abbreviation filter
          in: query
          required: false
          type: string
        - name: position
          description: Position abbreviation filter
          in: query
          required: false
          type: string
        - name: playerId
          description: Player id filter
          in: query
          required: false
          type: integer
        - name: excludeGarbageTime
          description: Filter to remove garbage time plays from calculations
          in: query
          required: false
          type: boolean
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/PlayerUsage'
        '400':
          description: error
  /player/returning:
    get:
      tags:
        - players
      summary: Team returning production metrics
      description: Returning production metrics
      operationId: getReturningProduction
      produces:
        - application/json
      parameters:
        - name: year
          description: Year filter
          in: query
          required: false
          type: integer
          minimum: 2014
        - name: team
          description: Team filter
          in: query
          required: false
          type: string
        - name: conference
          description: Conference abbreviation filter
          in: query
          required: false
          type: string
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/ReturningProduction'
        '400':
          description: error
...
  /player/portal:
    get:
      tags:
        - players
      summary: Transfer portal by season
      description: Transfer portal by season
      operationId: getTransferPortal
      produces:
        - application/json
      parameters:
        - name: year
          description: Year filter
          in: query
          required: true
          type: integer
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/PortalPlayer'
        '400':
          description: error
...
definitions:
...
  Team:
    type: object
    properties:
      id:
        type: integer
      school:
        type: string
      mascot:
        type: string
      abbreviation:
        type: string
      alt_name_1:
        type: string
      alt_name_2:
        type: string
      alt_name_3:
        type: string
      conference:
        type: string
      division:
        type: string
      color:
        type: string
      alt_color:
        type: string
      logos:
        type: array
        items:
          type: string
      location:
        type: object
        properties:
          venue_id:
            type: integer
          name:
            type: string
          city:
            type: string
          state:
            type: string
          zip:
            type: string
          country_code:
            type: string
          timezone:
            type: string
          latitude:
            type: number
          longitude:
            type: number
          elevation:
            type: number
          capacity:
            type: number
          year_constructed:
            type: number
          grass:
            type: boolean
          dome:
            type: boolean
...
  PlayerSearchResult:
    type: object
    properties:
      id:
        type: integer
      team:
        type: string
      name:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      weight:
        type: integer
      height:
        type: integer
      jersey:
        type: integer
      position:
        type: string
      hometown:
        type: string
      teamColor:
        type: string
      teamColorSecondary:
        type: string
...
  PlayerUsage:
    type: object
    properties: 
      season:
        type: integer
      id:
        type: integer
      name:
        type: string
      position:
        type: string
      team:
        type: string
      conference:
        type: string
      usage:
        type: object
        properties: 
          overall:
            type: number
          pass:
            type: number
          rush:
            type: number
          firstDown:
            type: number
          secondDown:
            type: number
          thirdDown:
            type: number
          standardDowns:
            type: number
          passingDowns:
            type: number
  ReturningProduction:
    type: object
    properties:
      season:
        type: integer
      team:
        type: string
      conference:
        type: string
      totalPPA:
        type: number
      totalPassingPPA:
        type: number
      totalReceivingPPA:
        type: number
      totalRushingPPA:
        type: number
      percentPPA:
        type: number
      percentPassingPPA:
        type: number
      percentReceivingPPA:
        type: number
      percentRushingPPA:
        type: number
      usage:
        type: number
      passingUsage:
        type: number
      receivingUsage:
        type: number
      rushingUsage:
        type: number
...
  PortalPlayer:
    type: object
    properties:
      season:
        type: integer
      firstName:
        type: string
      lastName:
        type: string
      position:
        type: string
      origin:
        type: string
      destination:
        type: string
      transferDate:
        type: string
      rating:
        type: number
      stars:
        type: integer
      eligibility:
        type: string